authorization implemented

+++ src/api/notes/handler.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 /* eslint-disable no-undef */
 /* eslint-disable no-underscore-dangle */
 
@@ -17,10 +18,15 @@ class NotesHandler {
 
   async postNoteHandler(request, h) {
     // try {
+
     this._validator.validateNotePayload(request.payload);
     const { title = 'untitled', body, tags } = request.payload;
+    const { id: credentialId } = request.auth.credentials;
 
-    const noteId = await this._service.addNote({ title, body, tags });
+    // const noteId = await this._service.addNote({ title, body, tags });
+    const noteId = await this._service.addNote({
+      title, body, tags, owner: credentialId,
+    });
 
     const response = h.response({
       status: 'success',
@@ -62,8 +68,9 @@ class NotesHandler {
     // }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -72,9 +79,12 @@ class NotesHandler {
     };
   }
 
-  async getNoteByIdHandler(request) {
+  async getNoteByIdHandler(request, h) {
     // try {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
     const note = await this._service.getNoteById(id);
     return {
       status: 'success',
@@ -115,10 +125,13 @@ class NotesHandler {
     // }
   }
 
-  async putNoteByIdHandler(request) {
+  async putNoteByIdHandler(request, h) {
     // try {
     this._validator.validateNotePayload(request.payload);
     const { id } = request.params;
+
+    const { id: credentialId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
 
     await this._service.editNoteById(id, request.payload);
 
@@ -159,9 +172,12 @@ class NotesHandler {
     // }
   }
 
-  async deleteNoteByIdHandler(request) {
+  async deleteNoteByIdHandler(request, h) {
     // try {
     const { id } = request.params;
+
+    const { id: credentialId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
 
     await this._service.deleteNoteById(id);
